+++ hibernate6-core/src/test/groovy/com/github/starnowski/posjsonhelper/hibernate6/descriptor/FunctionDescriptorRegisterFactoriesSupplierTest.groovy
@@ -22,7 +22,7 @@ class FunctionDescriptorRegisterFactoriesSupplierTest extends Specification {
             def results = tested.get()
 
         then:
-            results.stream().map({it -> it.getClass()}).collect(Collectors.toSet()) == new HashSet([TestFunctionDescriptorRegisterFactory1.class, TestFunctionDescriptorRegisterFactory2.class, CastOperatorFunctionDescriptorRegisterFactory.class])
+            results.stream().map({it -> it.getClass()}).collect(Collectors.toSet()) == new HashSet([TestFunctionDescriptorRegisterFactory1.class, TestFunctionDescriptorRegisterFactory2.class, CastOperatorFunctionDescriptorRegisterFactory.class, ArrayFunctionDescriptorRegisterFactory.class])
     }
 
     @Unroll