+++ hibernate6/src/main/java/com/github/starnowski/posjsonhelper/hibernate6/predicates/AbstractJsonbArrayStringsExistPredicate.java
@@ -55,7 +55,7 @@ public abstract class AbstractJsonbArrayStringsExistPredicate<T extends Abstract
 
     private final HibernateContext context;
     private final JsonBExtractPath jsonBExtractPath;
-    private final JsonArrayFunction jsonArrayFunction;
+    private final JsonArrayFunction<String> jsonArrayFunction;
 
     /**
      * @param context          object of type {@link HibernateContext}
@@ -75,7 +75,7 @@ public abstract class AbstractJsonbArrayStringsExistPredicate<T extends Abstract
      * @param arrayFunction    component of type {@link JsonArrayFunction} that represent array of values passed to as argument for function
      * @param functionName     function name
      */
-    public AbstractJsonbArrayStringsExistPredicate(HibernateContext context, NodeBuilder nodeBuilder, JsonBExtractPath jsonBExtractPath, JsonArrayFunction arrayFunction, String functionName) {
+    public AbstractJsonbArrayStringsExistPredicate(HibernateContext context, NodeBuilder nodeBuilder, JsonBExtractPath jsonBExtractPath, JsonArrayFunction<String> arrayFunction, String functionName) {
         super(nodeBuilder.getQueryEngine().getSqmFunctionRegistry().findFunctionDescriptor(functionName),
                 (FunctionRenderer) nodeBuilder.getQueryEngine().getSqmFunctionRegistry().findFunctionDescriptor(functionName),
                 parameters(jsonBExtractPath, arrayFunction),
@@ -89,21 +89,20 @@ public abstract class AbstractJsonbArrayStringsExistPredicate<T extends Abstract
         this.jsonArrayFunction = arrayFunction;
     }
 
-    private static List<? extends SqmExpression<String>> parameters(JsonBExtractPath jsonBExtractPath, JsonArrayFunction jsonArrayFunction) {
-        List<SqmExpression<String>> result = new ArrayList<>();
+    private static List<? extends SqmExpression<? extends String>> parameters(JsonBExtractPath jsonBExtractPath, JsonArrayFunction<String> jsonArrayFunction) {
+        List<SqmExpression<? extends String>> result = new ArrayList<>();
         result.add(jsonBExtractPath);
         result.add(jsonArrayFunction);
         return result;
     }
 
-    private static JsonArrayFunction mapArrayValues(NodeBuilder nodeBuilder, HibernateContext context, String... values) {
+    private static JsonArrayFunction<String> mapArrayValues(NodeBuilder nodeBuilder, HibernateContext context, String... values) {
         if (values == null || values.length == 0) {
             throw new IllegalArgumentException("Values can not be null or empty list");
         }
         List<SqmExpression<String>> arrayArguments = new ArrayList<>();
         arrayArguments.addAll(Stream.of(values).map(p -> (SqmExpression<String>) nodeBuilder.value(p)).collect(Collectors.toList()));
-        JsonArrayFunction jsonArrayFunction = new JsonArrayFunction(nodeBuilder, arrayArguments, context);
-        return jsonArrayFunction;
+        return new JsonArrayFunction<String>(nodeBuilder, arrayArguments, context);
     }
 
     public T copy(SqmCopyContext context) {
@@ -118,5 +117,5 @@ public abstract class AbstractJsonbArrayStringsExistPredicate<T extends Abstract
     }
 
 
-    abstract protected T generateCopy(HibernateContext context, NodeBuilder nodeBuilder, JsonBExtractPath jsonBExtractPath, JsonArrayFunction jsonArrayFunction);
+    abstract protected T generateCopy(HibernateContext context, NodeBuilder nodeBuilder, JsonBExtractPath jsonBExtractPath, JsonArrayFunction<String> jsonArrayFunction);
 }